Removing the assertion for exact pipeline label
In case of timer triggered pipeline with such short time as 30 secs, asserting for specific label of pipeline is not reliable, it causing intermittent test failure. The objective is to check if the pipeline is triggered by the timer or not. Which is validated by other steps. So removing this assertion of pipeline label