Fix for Stale Reference Exception
The scenario tested is that the next page is available while browsing the timeline for a given pipeline.
I suspect: the spec were getting executed before the capybara session could be updated for the given page- leading to a stale element getting referenced.
Fix: Adding a 1 sec delay before executing `page.find` currently ensures that the session has enough time to be able to reload