Try adding a wait kludge to "PipelineUsingTimerWithOnlyOnChangesOption"
For some reason after updating the Windows image, Chromium version etc, this test fails 80% of the time at the same step/verification. Sometimes it passes; but can't really figure out why or what makes it different to other similar assertions. Cannot replicate locally. The screenshot always shows the box disabled, as the test failure is indicating. Trying a kludge here to see whether it addresses it.

https://build.gocd.org/go/tab/build/detail/regression-SPAs/2404/Firefox/3/regression-runInstance-7#