Disable SystemAdminPrivilegeToggling which also is not working after changes to system admin
Users Management for system admins seems to have changed significantly, so this this was intentionally disabled it seems based on the comment below. Got it a little further in clicking the buttons properly, but it fails later on, and I can't easily figure out how it is supposed to work right now.