Workaround flaky test issues with Capybara "rapid" set_text
Since Capybara 3.32+ by default text setting for >30 chars will use a different, faster approach involving writing everything except the first 4 and last 3 characters directly via the driver. Unfortunately this seems to be a bit flaky in our setup (with either Selenium Webdriver 3.142.7 or 4.1.0) and the "fast typed" chars in the middle of a string can be completely missing from an input, for unknown reasons (possibly something in our Javascript?). Since this functionality isn't going away in Capybara (it's 18 months old now), we are better to disable it where required so we can move forward if we can't fix the root cause.

https://github.com/teamcapybara/capybara/blob/a958240ef44b44ca90be8e7b8fbcba8511a6e262/lib/capybara/selenium/node.rb#L277-L284
https://github.com/teamcapybara/capybara/issues/2521