Remove dependency on Jabba to determine Java version to run tests under
Prefer ASDF going forward since Jabba is unmaintained and doesn't support Adoptium/Temurin. We will work to this.

Will re-evaluate later if the scripting here should dynamically install a target Java version or not. Currently if the `.tool-versions` version is not available on the build agent it will fail. To decide if to use the system version, or to do a `asdf install`