Try cleaning working dir for "wait for file" stages
Not sure what is going on here, but some of these are re-triggered and expect the state to go to building. Some seem
to complete too fast or unexpectedly, so possible the file is there from a previous run if outside the git/clone dir
(which should be cleaned) and inside the working dir (which is not specifically cleaned by default)?