Remove flakey verifications
It's really not clear why these verifications often fail. Screenshots indicate the output is as expected, so there must
be some CSS selector or timing issue. Instead of directly verifying the group visibility, use one of the pipelines inside
the group to validate and see if we can increase stability.