Remove EFS logic
It's not necessary to specifically test on EFS, it's just a shared file system and this actually specifically tests the ECS elastic plugin logic
but in a very expensive way - without actually being able to tell the underlying file system. This test passes if there is any docker mount whether it's a real network filesystem or not.