Try disabling agents before destroying them, and then deleting them.
Sometimes we seem to try and delete an agent too quickly, before it has fully completed its teardown. Maybe disabling it while it's still up will be better?  Might also be better to delete agents one-by-one and check the response so we can retry if the agent still says it is in building state (which prevents deletion)